ENH: Initial itkHalideDiscreteGaussianImageFilter (#10)
* WIP: ENH: Add trivial halide cast image filter

* WIP: FIX: Use GenerateData() rather than DynamicThreadedGenerateData()

* WIP: Simple itkHalideDiscreteGaussianImpl generator stub; integrated and working.

* WIP: ENH: Implement HalideDiscreteGaussian with naive schedule.

Implemented with separable convolution. Handles separate sigma and kernel size for each dimension to support anisotropic images.

Strip out the CastImageFilter and MinimalStandardRandomVariateGenerator placeholders.

* WIP: ENH: Add naive CUDA schedule

* ENH: Improved CUDA schedule.

About 200ms on my machine. Bandwidth-limited on reasonable kernel/image sizes.

* ENH: Generate kernels with itk::GaussianOperator

for parity with itk::DiscreteGaussianImageFilter.

* FIX: Disable multithreading.

It wasn't used anyway, but this should make the filter better-behaved.

Not sure how to disable/remove the MultiThreader object; setting it null is not valid.

* FIX: Verification Test

Change the test to blur CTChest. The reference output was generated with itk::DiscreteGaussianImageFilter directly, so the Halide output is verified to be numerically compatible.

---------

Co-authored-by: Anthony Lombardi <anthony.lombardi@kitware.com>